Rewrite core sim.c for threadsafety and reentrant
With the new threadsafe primatives, I have reimplmeneted
the old threadunsafe API functions.  Tested against pmars
and old versions of exhaust.

The core of the change is to operate of SimState_t objects
rather than a bunch of gross static globals.  Legacy API
passes a static global SimState_t object to the primatives