Remove PSPACE functionality from the stripped branch
PSPACE is not relevent and causes ~5% overhead to evolver benchmarks

It isn't used in the compition that I am evolving for, and it
causes a great deal of mental overhead to deal with.

The pspace code always felt a bit bolted on as an afterthought,
and the allocation/freeing/clearing of a seperate memory bank
clutters things up.  I'm glad to be rid of it on the stripped
branch.